fix: add prop into requestDto

+++ AptCare.Service/Dtos/IssueDto/IssueListItemDto.cs
@@ -15,7 +15,7 @@ namespace AptCare.Service.Dtos.IssueDto
         public string Name { get; set; } = default!;
         public bool IsEmergency { get; set; }
         public int RequiredTechnician { get; set; }
-        public int EstimatedDuration { get; set; }
+        public double EstimatedDuration { get; set; }
         public string Status { get; set; } = "Active";
     }
 }
